split test file to go1.22
